+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes qsf-integration project will be documented in this file.
 
 ## [UNRELEASED]
 ## [3.1]
+### Added
+- add mapper for SearchQuery and SearchQueryDTO
+
 ### Changed
 - use qsf-commons dependency
 - update jackson-dataformat-yaml to 2.10.3